Split export tests per format and cover query forwarding
Split the DatasetExport format tests into per-format files
(test_dataset_export__coco_object_detections.py, __yolo_object_detections.py,
__coco_captions.py, __coco_segmentation_masks.py, __pascal_voc.py), matching the
test_<subject>__<topic>.py convention used in tests/core/image. The shared
DatasetExport factory moves to tests/export/helpers.py.

Restore query-forwarding coverage in the ImageDataset.export() smoke test, and
drop the "bind"/"substrate" wording from docstrings.
